+++ R/forest.netmeta.R
@@ -551,7 +551,7 @@ forest.netmeta <- function(x,
                                    warn = FALSE))
   else
     m1 <- suppressWarnings(metagen(TE, seTE, data = dat,
-                                   byvar = dat$comparison,
+                                   subgroup = dat$comparison,
                                    sm = x$sm,
                                    studlab = labels, backtransf = backtransf,
                                    method.tau = "DL", method.tau.ci = "",